requirements.txt
asgiref
backports.zoneinfo
Django
django-ckeditor
django-js-asset
Pillow
sqlparse
whitenoise
